+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2025-07-18
+
+### Added
+- `widgets.ColoredButton`: as the name suggests ...
+
+### Changed
+- `widgets.SlideToggle`: tweaked color-scheme
+
 ## [0.7.1] - 2025-07-16
 
 ### Changed
@@ -145,6 +153,7 @@ _First release._
   providing color-mapped numerical data.
 - `widgets.StatefulButton`: A `QPushButton` that maintains an active/inactive state.
 
+[0.8.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.8.0
 [0.7.1]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.7.1
 [0.7.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.7.0
 [0.6.1]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.6.1